[DLXEmu] Change `m_TabSize` type to std::uint_fast8_t

+++ DLXEmu/include/DLXEmu/CodeEditor.hpp
@@ -31,6 +31,7 @@ SOFTWARE.
 #include <Phi/Core/Types.hpp>
 #include <imgui.h>
 #include <array>
+#include <cstdint>
 #include <map>
 #include <string>
 #include <unordered_map>
@@ -312,21 +313,21 @@ namespace dlxemu
         UndoBuffer  m_UndoBuffer;
         std::size_t m_UndoIndex;
 
-        std::uint32_t m_TabSize;
-        bool          m_Overwrite : 1;
-        bool          m_ReadOnly : 1;
-        bool          m_WithinRender : 1;
-        bool          m_ScrollToCursor : 1;
-        bool          m_ScrollToTop : 1;
-        bool          m_TextChanged : 1;
-        bool          m_ColorizerEnabled : 1;
-        bool          m_CursorPositionChanged : 1;
-        float         m_TextStart; // position (in pixels) where a code line starts relative to the left of the CodeEditor.
-        std::int32_t  m_LeftMargin;
-        std::int32_t  m_ColorRangeMin;
-        std::int32_t  m_ColorRangeMax;
-        SelectionMode m_SelectionMode;
-        bool          m_ShowWhitespaces : 1;
+        std::uint_fast8_t m_TabSize;
+        bool              m_Overwrite : 1;
+        bool              m_ReadOnly : 1;
+        bool              m_WithinRender : 1;
+        bool              m_ScrollToCursor : 1;
+        bool              m_ScrollToTop : 1;
+        bool              m_TextChanged : 1;
+        bool              m_ColorizerEnabled : 1;
+        bool              m_CursorPositionChanged : 1;
+        float             m_TextStart; // position (in pixels) where a code line starts relative to the left of the CodeEditor.
+        std::int32_t      m_LeftMargin;
+        std::int32_t      m_ColorRangeMin;
+        std::int32_t      m_ColorRangeMax;
+        SelectionMode     m_SelectionMode;
+        bool              m_ShowWhitespaces : 1;
 
         Palette m_PaletteBase;
         Palette m_Palette;
